more blog and page stuff plus some refactoring

+++ app/Blog/BlogPost/EloquentBlogPostRepository.php
@@ -37,42 +37,73 @@ class EloquentBlogPostRepository implements BlogPostRepositoryInterface {
     }
 
     public function updateBlogPost($input) {
-        $this->blogPost = $this->blogPost->where('id', $input['id'])->first();
-        $this->blogPost->author_id = $input['author'];
-        $this->blogPost->title = $input['title'];
-        $this->blogPost->intro_text = $input['intro_text'];
-        $this->blogPost->body_text = $input['body_text'];
-        $this->blogPost->updated_by = $this->auth->user()->getAuthIdentifier();
-        $this->blogPost->url = $input['url'];
-        $this->blogPost->meta_title = $input['meta_title'];
-        $this->blogPost->meta_keywords = $input['meta_keywords'];
-        $this->blogPost->meta_description = $input['meta_description'];
-        $this->blogPost->save();
-        $this->blogPost->categories()->sync($input['categories']);
-        return $this->blogPost->toArray();
+        $this->blogPost = $this->blogPost
+            ->where('id', $input['id'])
+            ->first();
+        if (!is_null($this->blogPost)) {
+            $this->blogPost->author_id = $input['author'];
+            $this->blogPost->title = $input['title'];
+            $this->blogPost->intro_text = $input['intro_text'];
+            $this->blogPost->body_text = $input['body_text'];
+            $this->blogPost->updated_by = $this->auth->user()->getAuthIdentifier();
+            $this->blogPost->url = $input['url'];
+            $this->blogPost->meta_title = $input['meta_title'];
+            $this->blogPost->meta_keywords = $input['meta_keywords'];
+            $this->blogPost->meta_description = $input['meta_description'];
+            $this->blogPost->save();
+            $this->blogPost->categories()->sync($input['categories']);
+            $this->blogPost = $this->blogPost->toArray();
+        }
+        return $this->blogPost;
     }
 
     public function deleteBlogPost($id) {
-        $this->blogPost = $this->blogPost->where('id', $id)->first();
-        return $this->blogPost->delete();
+        $this->blogPost = $this->blogPost
+            ->where('id', $id)
+            ->first();
+        if (!is_null($this->blogPost)) {
+            $this->blogPost->delete();
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public function publishBlogPost($id) {
-        $this->blogPost = $this->blogPost->where('id', $id)->first();
-        $this->blogPost->enabled = self::ENABLED;
-        return $this->blogPost->save();
+        $this->blogPost = $this->blogPost
+            ->where('id', $id)
+            ->first();
+        if (!is_null($this->blogPost)) {
+            $this->blogPost->enabled = self::ENABLED;
+            $this->blogPost->save();
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public function unpublishBlogPost($id) {
-        $this->blogPost = $this->blogPost->where('id', $id)->first();
-        $this->blogPost->enabled = self::DISABLED;
-        return $this->blogPost->save();
+        $this->blogPost = $this->blogPost
+            ->where('id', $id)
+            ->first();
+        if (!is_null($this->blogPost)) {
+            $this->blogPost->enabled = self::DISABLED;
+            $this->blogPost->save();
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public function getBlogPostById($id) {
         $post = $this->blogPost
             ->where('id', $id)
-            ->with(['author', 'author.social', 'comments', 'comments.author', 'categories'])
+            ->with([
+                'author',
+                'author.social',
+                'comments',
+                'comments.author',
+                'categories'])
             ->first();
         if (!is_null($post)) {
             $post = $post->toArray();
@@ -88,12 +119,20 @@ class EloquentBlogPostRepository implements BlogPostRepositoryInterface {
         if ($enabled_only) {
             $posts = $this->blogPost
                 ->where('enabled', self::ENABLED)
-                ->with(['author', 'comments', 'categories'])
+                ->with([
+                    'author',
+                    'comments',
+                    'categories'
+                ])
                 ->orderBy('created_at', 'desc')
                 ->paginate(self::POSTS_PER_PAGE_BLOG);
         } else {
             $posts = $this->blogPost
-                ->with(['author', 'comments', 'categories'])
+                ->with([
+                    'author',
+                    'comments',
+                    'categories'
+                ])
                 ->orderBy('created_at', 'desc')
                 ->paginate(self::POSTS_PER_PAGE_ADMIN);
         }
